working version of CK Editor - without documentation or tests
