+++ README.md
@@ -1,2 +1,11 @@
-# CkHelper
-CK Helper for CakePHP 3.x
+# CkEditor plugin for CakePHP
+
+## Installation
+
+You can install this plugin into your CakePHP application using [composer](http://getcomposer.org).
+
+The recommended way to install composer packages is:
+
+```
+composer require CakeCoded/CkEditor
+```
