+++ composer.json
@@ -1,21 +1,13 @@
 {
     "name": "cakecoded/ckeditor",
-    "description": "CKEditor Helper Plugin for CakePHP 3.x.",
-    "homepage": "http://github.com/CakeCoded/CkHelper",
+    "description": "CkEditor plugin for CakePHP 3.x",
     "type": "cakephp-plugin",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "CakeCoded",
-            "homepage": "http://www.cakecoded.com"
-        }
-    ],
     "require": {
-        "php": ">=5.4.16",
-        "cakephp/cakephp": "~3.0"
+        "php": ">=5.5.9",
+        "cakephp/cakephp": ">=3.3.2 <4.0.0"
     },
     "require-dev": {
-        "composer/composer": "1.0.*@dev"
+        "phpunit/phpunit": "*"
     },
     "autoload": {
         "psr-4": {
@@ -24,8 +16,8 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Cake\\Test\\": "vendor/cakephp/cakephp/tests",
+            "CkEditor\\Test\\": "tests",
+            "Cake\\Test\\": "./vendor/cakephp/cakephp/tests"
         }
-    },
-    "minimum-stability": "dev"
+    }
 }
